+++ src/nebulib/purplenebula/ui/enums/ProgressBarLabelType.java
@@ -1,0 +1,7 @@
+package nebulib.purplenebula.ui.enums;
+
+public enum ProgressBarLabelType {
+    PERCENTAGE,
+    CURRENT_TOTAL,
+    CURRENT,
+}