+++ src/nebulib/purplenebula/ui/enums/ProgressBarType.java
@@ -1,0 +1,8 @@
+package nebulib.purplenebula.ui.enums;
+
+public enum ProgressBarType {
+    GOOD_BAD,
+    REL,
+    LOYALTY,
+    DEFAULT
+}