+++ src/nebulib/purplenebula/utils/StarTypes.java
@@ -1,4 +1,4 @@
-package com.purplenebula.nebulib.util;
+package nebulib.purplenebula.utils;
 
 /**
  * @author Purple Nebula (NebulaModLab)