+++ CHANGELOG.md
@@ -1,10 +1,18 @@
+Version 1.2.0
+
+**Changes**
+- Added: ProgressBarComponent class, allowing easy creation of progression bars of different types.
+  - Thank you Kaysaar for improving my original code for this
+- Added: ProgressBarType enum for the ProgressBarComponent
+- Added: ProgressBarLabelType enum for the ProgressBarComponent
+
 Version 1.1.0
 
-**NebuLib**
+**Changes**
 - Initial release of NebuLib even though it says v1.1.0 - this is due to a name refactoring of the entire mod.
 
 
 Version 1.0.0
 
-**NebuLib**
+**Changes**
 - Old initial release under and old name.